requirements.txt
instagrapi>=1.19.6
python-dotenv>=1.0.1
google-genai>=0.1.0
